more msvc warning fixes

+++ docs/code-examples/depth_image_simple.cpp
@@ -13,10 +13,10 @@ int main() {
     const int WIDTH = 300;
     std::vector<uint16_t> data(WIDTH * HEIGHT, 65535);
     for (auto y = 50; y < 150; ++y) {
-        std::fill_n(data.begin() + y * WIDTH + 50, 100, 20000);
+        std::fill_n(data.begin() + y * WIDTH + 50, 100, static_cast<uint16_t>(20000));
     }
     for (auto y = 130; y < 180; ++y) {
-        std::fill_n(data.begin() + y * WIDTH + 100, 180, 45000);
+        std::fill_n(data.begin() + y * WIDTH + 100, 180, static_cast<uint16_t>(45000));
     }
 
     rec.log("depth", rerun::DepthImage({HEIGHT, WIDTH}, std::move(data)).with_meter(10000.0));
